# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Order Management API
  version: 1.0.0
  description: API для управления заказами
servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки

paths:
  /orders/create:
    post:
      tags:
        - Orders
      summary: Создать новый заказ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrdersEntity'
      responses:
        '200':
          description: Заказ успешно создан
          content:
            text/plain:
              schema:
                type: string
                example: "Success"

  /orders/find/{id}:
    get:
      tags:
        - Orders
      summary: Получить заказ по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Найденный заказ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrdersDto'
        '404':
          description: Заказ не найден

  /orders/findAll:
    get:
      tags:
        - Orders
      summary: Получить заказы по дате и минимальной сумме
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
            pattern: '^\d{8}$'
            example: "20250128"
        - name: minTotal
          in: query
          required: true
          schema:
            type: integer
            example: 100
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrdersDto'

  /orders/without-product:
    get:
      tags:
        - Orders
      summary: Получить заказы без указанного товара
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
            pattern: '^\d{8}$'
            example: "20250128"
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
            pattern: '^\d{8}$'
            example: "20250131"
        - name: articleProduct
          in: query
          required: true
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Список заказов без указанного товара
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrdersDtoWithoutProduct'

  /orders/{key}:
    get:
      tags:
        - Orders
      summary: Получить номер заказа по ключу
      description: Возвращает заказ по уникальному идентификатору из Redis
      parameters:
        - name: key
          in: path
          required: true
          description: Уникальный идентификатор заказа (ключ в Redis)
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Найден заказ
          content:
            text/plain:
              schema:
                type: string
                example: "ORD-12345"
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Номер не найден"
        '500':
          description: Ошибка сервера
          content:
            application.json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"

components:
  schemas:
    OrdersEntity:
      type: object
      properties:
        orderNumber:
          type: string
          example: "ORD-123"
        totalOrderAmount:
          type: integer
          example: 1000
        orderDate:
          type: string
          format: date
          example: "2025-01-28"
        address:
          type: string
          example: "ул. Тестовая, 1"
        deliveryAddress:
          type: string
          example: "ул. Доставки, 5"
        paymentType:
          type: string
          example: "CARD"
        deliveryType:
          type: string
          example: "COURIER"
        ordersDetails:
          type: array
          items:
            $ref: '#/components/schemas/OrdersDetailsEntity'

    OrdersDetailsEntity:
      type: object
      properties:
        articleProduct:
          type: integer
          example: 1001
        nameProduct:
          type: string
          example: "Тестовый товар 1"
        quantity:
          type: integer
          example: 2
        unitCost:
          type: integer
          example: 300

    OrdersDto:
      type: object
      properties:
        id:
          type: integer
          example: 1
        orderNumber:
          type: string
          example: "ORD-123"
        totalOrderAmount:
          type: integer
          example: 1000
        orderDate:
          type: string
          format: date
          example: "2025-01-28"
        address:
          type: string
          example: "ул. Тестовая, 1"
        deliveryAddress:
          type: string
          example: "ул. Доставки, 5"
        paymentType:
          type: string
          example: "CARD"
        deliveryType:
          type: string
          example: "COURIER"

    OrdersDtoWithoutProduct:
      type: object
      properties:
        id:
          type: integer
          example: 1
        orderNumber:
          type: string
          example: "ORD-123"
        totalOrderAmount:
          type: integer
          example: 1000
        orderDate:
          type: string
          format: date
          example: "2025-01-28"
        address:
          type: string
          example: "ул. Тестовая, 1"
        deliveryAddress:
          type: string
          example: "ул. Доставки, 5"
        paymentType:
          type: string
          example: "CARD"
        deliveryType:
          type: string
          example: "COURIER"